+++ CHANGELOG.md
@@ -6,3 +6,13 @@
 ## [0.0.1+1] - 2021-05-12
 
 * Updated README and documentation
+
+## [0.0.2] - 2021-05-18
+
+* Updated all the Firebase Dependencies to a combination that works
+* FireAuth Phone Authentication: Mobile onError Callbacks work
+* FireAuth Phone Authentication: VerificationDialog closes on verificationCompleted
+* Generic Code Improvements
+* Added a success callback to all AuthController static methods
+* Removed AndroidX BrowserImplementation in app level build.gradle (example app)
+* (example > main.dart) Added success callback examples